requirements.txt
wheel
PyYaml
numpy
protobuf
tensorflow